+++ .gitignore
@@ -1,0 +1,5 @@
+
+.classpath
+.project
+bin/
+*.prefs
