--- bake/.classpath
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<classpath>
-	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7">
-		<attributes>
-			<attribute name="module" value="true"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="/api/spigot-api-1.13.2-R0.1-20190423.030901-214.jar" sourcepath="C:/Users/EmericW/Desktop/Geolykt/mc/bukkit/api/spigot-api-1.13.2-R0.1-20190423.030901-214-sources.jar">
-		<attributes>
-			<attribute name="javadoc_location" value="jar:file:/api/spigot-api-1.13.2-R0.1-20190423.030901-214-javadoc.jar!/"/>
-		</attributes>
-	</classpathentry>
-	<classpathentry kind="output" path="bin"/>
-</classpath>
